Rename components to .js extension
