+++ src/assets/img/icons/favicons.js
@@ -1,9 +1,3 @@
-import favicon16 from './favicon-16.png';
-import favicon32 from './favicon-32.png';
-import favicon96 from './favicon-96.png';
-
-export default {
-  favicon16,
-  favicon32,
-  favicon96,
-};
+export { default as favicon16 } from './favicon-16.png';
+export { default as favicon32 } from './favicon-32.png';
+export { default as favicon96 } from './favicon-96.png';
