+++ src/components/comments/index.js
@@ -1,0 +1,1 @@
+export { default as Comments } from './Comments';
