+++ src/components/footer/index.js
@@ -1,0 +1,2 @@
+export { default as Pagination } from './Pagination';
+export { default as Footer } from './Footer';
