+++ src/components/helmet/index.js
@@ -1,0 +1,3 @@
+export { default as Helmet } from './Head';
+export { default as PageHelmet } from './PageHead';
+export { default as PostHelmet } from './PostHead';
