+++ src/components/post/Tags.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { array } from 'prop-types';
 import Link from 'gatsby-link';
 
-import './styles/tag.css';
+import './styles/Tags.css';
 
 // TODO: Add a link which goes to the tag-group page
 const Tags = ({ tags }) => (
