+++ src/components/post/index.js
@@ -1,0 +1,3 @@
+export { default as ContinueReadingButton } from './ContinueReadingButton';
+export { default as Preview } from './Preview';
+export { default as Tags } from './Tags';
