+++ src/components/recent-repos/RecentRepos.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import fetch from 'isomorphic-fetch';
 
-import Repo from './Repo';
+import { Repo } from './';
 
 import './index.css';
