+++ src/components/recent-repos/index.js
@@ -1,0 +1,2 @@
+export { default as RecentRepos } from './RecentRepos';
+export { default as Repo } from './Repo';
