+++ src/pages/Currents.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { object } from 'prop-types';
 
-import Helmet from '../components/helmet/page';
+import { PageHelmet } from '../components/helmet/';
 
 const Currents = ({ data }) => {
   const { title, description, list } = data.contentfulList;
 
   return (
     <section>
-      <Helmet title={title} />
+      <PageHelmet title={title} />
       <header>
         <h1>{title}</h1>
       </header>
