+++ src/pages/Uses.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { object } from 'prop-types';
 
-import Helmet from '../components/helmet/page';
+import { PageHelmet } from '../components/helmet/';
 
 const Publications = ({ data }) => {
   const { title, body } = data.contentfulPage;
 
   return (
     <section>
-      <Helmet title={title} />
+      <PageHelmet title={title} />
       <header>
         <h1>{title}</h1>
       </header>
